+++ src/Ampersand/Event/LinkEvent.php
@@ -4,6 +4,7 @@ namespace Ampersand\Event;
 
 use Ampersand\Core\Link;
 use Ampersand\Event\AbstractEvent;
+use Ampersand\Transaction;
 
 class LinkEvent extends AbstractEvent
 {
@@ -12,14 +13,26 @@ class LinkEvent extends AbstractEvent
         DELETED = 'ampersand.core.link.deleted';
 
     protected Link $link;
+    protected Transaction $transaction;
 
-    public function __construct(Link $link)
+    public function __construct(Link $link, Transaction $transaction)
     {
         $this->link = $link;
+        $this->transaction = $transaction;
     }
 
     public function getLink(): Link
     {
         return $this->link;
+    }
+
+    public function getTransaction(): Transaction
+    {
+        return $this->transaction;
+    }
+
+    public function isCommitted(): bool
+    {
+        return $this->transaction->isCommitted();
     }
 }